requirements.txt
pandas==2.2.3
requests==2.32.3
beautifulsoup4==4.13.3
python-dotenv==1.1.0
Jinja2==3.1.6
sendgrid==6.11.0
boto3==1.37.33
botocore==1.37.33